Rain_Cloud_Tracking_Firmware/requirements.txt
Adafruit-Blinka==8.35.0
adafruit-circuitpython-bme280==2.6.24
adafruit-circuitpython-busdevice==5.2.6
adafruit-circuitpython-connectionmanager==1.0.1
adafruit-circuitpython-requests==3.2.0
adafruit-circuitpython-typing==1.10.3
Adafruit-PlatformDetect==3.62.0
Adafruit-PureIO==1.1.11
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Faker==25.0.0
Flask==3.0.2
geographiclib==2.0
geopy==2.4.1
idna==3.7
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mpu6050-raspberrypi==1.2
pyftdi==0.55.0
pynmea2==1.19.0
pyserial==3.5
python-dateutil==2.9.0.post0
pyusb==1.2.1
requests==2.31.0
rpi-ws281x==5.0.0
RPi.GPIO==0.7.1
six==1.16.0
smbus==1.1.post2
sysv-ipc==1.1.0
typing_extensions==4.10.0
urllib3==2.2.1
Werkzeug==3.0.1
